🍫 (#95) - change file name and change queries code
